+++ src/extension.ts
@@ -1,35 +1,53 @@
-import * as vscode from 'vscode';
-import { TextDocument, window, workspace } from 'vscode';
+import {
+  commands,
+  ConfigurationChangeEvent,
+  ExtensionContext,
+  TextDocument,
+  TextDocumentChangeEvent,
+  TextEditor,
+  window,
+  workspace
+} from 'vscode';
 import { Commands } from './models';
 import { estimateReadTime, toggleEnableHandler } from './commands';
 import { Logger } from './logging';
+import { getEnabledSetting } from './configuration';
 
-export function activate(context: vscode.ExtensionContext) {
-  context.subscriptions.push(
-    vscode.commands.registerCommand(Commands.toggleEnable, toggleEnableHandler)
+export function activate(context: ExtensionContext) {
+  const { subscriptions } = context;
+  subscriptions.push(
+    commands.registerCommand(Commands.toggleEnable, toggleEnableHandler)
   );
 
   addSubscriptions(context);
 }
 
-function addSubscriptions(context: vscode.ExtensionContext) {
-  context.subscriptions.push(Logger.getChannel());
+function addSubscriptions({ subscriptions }: ExtensionContext) {
+  subscriptions.push(Logger.getChannel());
 
-  context.subscriptions.push(
-    workspace.onDidChangeConfiguration(estimateReadTime)
-  );
-
-  context.subscriptions.push(
-    window.onDidChangeActiveTextEditor((e: vscode.TextEditor | undefined) => {
+  subscriptions.push(
+    workspace.onDidChangeConfiguration((e: ConfigurationChangeEvent) => {
+      const isEnabled = getEnabledSetting();
+      Logger.info(
+        `Configuration changed. Read Time is now ${
+          isEnabled ? 'enabled' : 'disabled'
+        }`
+      );
       estimateReadTime();
     })
   );
-  context.subscriptions.push(
-    workspace.onDidChangeTextDocument((e: vscode.TextDocumentChangeEvent) => {
+
+  subscriptions.push(
+    window.onDidChangeActiveTextEditor((e: TextEditor | undefined) => {
       estimateReadTime();
     })
   );
-  context.subscriptions.push(
+  subscriptions.push(
+    workspace.onDidChangeTextDocument((e: TextDocumentChangeEvent) => {
+      estimateReadTime();
+    })
+  );
+  subscriptions.push(
     workspace.onDidOpenTextDocument((document: TextDocument) => {
       estimateReadTime();
     })