+++ src/models/enums.ts
@@ -1,11 +1,13 @@
 export enum Commands {
-  toggleEnable = 'read-time.toggleEnable'
+  toggleEnable = 'readTime.toggleEnable'
 }
 
 export enum Sections {
-  userReadTimeSection = 'read-time'
+  userReadTimeSection = 'readTime'
 }
 
 export enum Settings {
-  Enabled = 'enabled'
+  Enabled = 'enabled',
+  FileTypes = 'fileTypes',
+  WordsPerMinute = 'wordsPerMinute'
 }