+++ src/models/interfaces.ts
@@ -16,5 +16,14 @@ export interface IConfiguration {
 
 export interface IReadTimeSettings {
   enabled: boolean;
-  delay: number;
+  fileTypes: string[];
+  wordsPerMinute: number;
+}
+
+export interface IExpandedReadingTimeData {
+  roundedMinutes: number;
+  text: string;
+  time: number;
+  words: number;
+  minutes: number;
 }