+++ src/readtime.ts
@@ -1,17 +1,19 @@
 import * as readingTime from 'reading-time';
-import { TextDocument } from 'vscode';
-import { ReadingTimeData } from './models';
+import { ReadingTimeData, IExpandedReadingTimeData } from './models';
+import { getWPMSetting } from './configuration';
 
-export function getReadingTime(document: TextDocument) {
-  //textToRead: string) {
+export function getReadingTime(textToRead: string): IExpandedReadingTimeData {
   let readingTimeData = new ReadingTimeData();
-  if (document.languageId === 'markdown') {
-    const textToRead = document.getText();
-    // const options = {
-    //   wordBound: () => {},
-    //   wordsPerMinute: 200
-    // };
-    readingTimeData = readingTime(textToRead); //, options); // TODO: API for npm package is not up to date.
-  }
-  return readingTimeData;
+  const wordsPerMinute = getWPMSetting();
+  const options = { wordsPerMinute };
+  readingTimeData = readingTime(textToRead, options);
+  return modifyReadingData(readingTimeData);
+}
+
+function modifyReadingData(
+  readingTimeData: ReadingTimeData
+): IExpandedReadingTimeData {
+  const minutes = Math.round(readingTimeData.minutes);
+  const text = `${minutes} minute read`;
+  return { ...readingTimeData, roundedMinutes: minutes, text };
 }