+++ CHANGELOG.md
@@ -4,4 +4,11 @@ All notable changes to the code will be documented in this file.
 
 ## 1.0.0
 
-- Launch!
+Features
+
+- Create `wordsPerMinute` setting that defaults to 200
+- Created the `fileTypes` setting to indicate which files to estimate time to read
+
+Bug Fixes
+
+- Hide the status bar icon when not in a markdown file